refactor: condense project into three modules
