+++ requirements.txt
@@ -1,0 +1,6 @@
+python-telegram-bot
+python-dotenv
+torch
+numpy
+tiktoken
+PyYAML
